requirements.txt
# Core data handling
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0      # Excel support

# Document parsing
pdfplumber>=0.10.0   # PDF extraction (tables + text)
PyMuPDF>=1.23.0      # PDF extraction (fast, text-focused)
python-pptx>=0.6.21  # PowerPoint extraction
python-docx>=0.8.11  # Word extraction

# Image / OCR processing
Pillow>=10.0.0       # Image handling
easyocr>=1.7.0       # OCR — pure Python, no system packages needed
# Optional (faster OCR if tesseract-ocr is installed at OS level):
# pytesseract>=0.3.10
# opencv-python-headless>=4.8.0  # Better image preprocessing (optional, Pillow used as fallback)

# Visualization (for later phases)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistics (for later phases)
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# AI integration (optional — choose one or none)
# ollama>=0.1.0        # Ollama client (requires Ollama server running: https://ollama.com)
# llama-cpp-python>=0.2.0  # Local GGUF inference, no server needed (pip install llama-cpp-python)
#   → Download a GGUF model, e.g. Phi-3-mini from https://huggingface.co/microsoft/Phi-3-mini-4k-instruct-gguf

# Configuration
pyyaml>=6.0

# UI
streamlit>=1.28.0

# Development
pytest>=7.4.0
